// src/enums/error-code.enum.ts
export enum ErrorCode {
  PROMO_CODE_NOT_FOUND = 'PROMO_CODE_NOT_FOUND',
  PROMO_CODE_EXISTS = 'PROMO_CODE_EXISTS',
  // VALIDATION_ERROR = 'VALIDATION_ERROR',
  PROMO_CODE_EXPIRED = 'PROMO_CODE_EXPIRED',
  ACTIVATION_LIMIT_REACHED = 'ACTIVATION_LIMIT_REACHED',
  ALREADY_ACTIVATED = 'ALREADY_ACTIVATED',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
  // NOT_FOUND = 'NOT_FOUND',
}